Fix bug in event update REST endpoint
Prior to this commit, PUT requests for events actually saved a new event
instead of updating the existing one. The reason is that Spring’s
WebDataBinder is configured to disallow binding for all ‘id’ fields.
This is done for security reasons and therefore should not be changed.

Consequently, the updateEvent() REST endpoint now uses Spring’s
DirectFieldAccessor to set the updated event’s ID via reflection.